:poop: add & modify base crud files by bootify

+++ src/main/java/spring/grepp/honlife/app/controller/badge/BadgeController.java
@@ -36,27 +36,27 @@ public class BadgeController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<BadgeDTO> getBadge(@PathVariable(name = "id") final Integer id) {
+    public ResponseEntity<BadgeDTO> getBadge(@PathVariable(name = "id") final Long id) {
         return ResponseEntity.ok(badgeService.get(id));
     }
 
     @PostMapping
     @ApiResponse(responseCode = "201")
-    public ResponseEntity<Integer> createBadge(@RequestBody @Valid final BadgeDTO badgeDTO) {
-        final Integer createdId = badgeService.create(badgeDTO);
+    public ResponseEntity<Long> createBadge(@RequestBody @Valid final BadgeDTO badgeDTO) {
+        final Long createdId = badgeService.create(badgeDTO);
         return new ResponseEntity<>(createdId, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Integer> updateBadge(@PathVariable(name = "id") final Integer id,
-            @RequestBody @Valid final BadgeDTO badgeDTO) {
+    public ResponseEntity<Long> updateBadge(@PathVariable(name = "id") final Long id,
+        @RequestBody @Valid final BadgeDTO badgeDTO) {
         badgeService.update(id, badgeDTO);
         return ResponseEntity.ok(id);
     }
 
     @DeleteMapping("/{id}")
     @ApiResponse(responseCode = "204")
-    public ResponseEntity<Void> deleteBadge(@PathVariable(name = "id") final Integer id) {
+    public ResponseEntity<Void> deleteBadge(@PathVariable(name = "id") final Long id) {
         final ReferencedWarning referencedWarning = badgeService.getReferencedWarning(id);
         if (referencedWarning != null) {
             throw new ReferencedException(referencedWarning);
