+++ src/main/java/spring/grepp/honlife/app/model/notification/repos/NotificationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import spring.grepp.honlife.app.model.notification.domain.Notification;
 
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
 }
