+++ src/main/java/spring/grepp/honlife/app/model/routine/repos/RoutinePresetRepository.java
@@ -5,7 +5,7 @@ import spring.grepp.honlife.app.model.category.domain.Category;
 import spring.grepp.honlife.app.model.routine.domain.RoutinePreset;
 
 
-public interface RoutinePresetRepository extends JpaRepository<RoutinePreset, Integer> {
+public interface RoutinePresetRepository extends JpaRepository<RoutinePreset, Long> {
 
     RoutinePreset findFirstByCategory(Category category);
